src/restful/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0,<4.0
Markdown>=3.0,<4.0
django-filter>=2.0,<3.0
